Switch user shell to zsh
While the shell configuration is now placed withing the usershell task
file, I don't find it ideal. On one hand the shell configuration does
not need to be repo, could also be contained a dotfiles repo. On the
other hand the usershell task should only care about changing a users
shell not about configuring it.

TODO: Leave it for now where it is and revisit when it's clear how and
where to manage such configurations.